feat: consolidate MessageInput with @mention support
- Merge TeamChatMessageInput into MessageInput with optional mention support
- Add mentionableMembers prop to enable @mention autocomplete
- Only show dropdown at word boundaries (prevents emails like tcm390@nyu.edu)
- Add highlight overlay for valid @mentions in input
- Update Response.tsx to only format @username if in validMentions list
- Pass validMentions through MessageBubble and MessageList
- Add username to OwnedAgentData and AgentChat props
- Remove TeamChatMessageInput.tsx (consolidated into MessageInput)
- Update ChatView, AgentChat, TeamChatView to use consolidated component
