+++ apps/web/src/components/chats/index.ts
@@ -14,6 +14,5 @@ export {
 export { MessageBubble } from './MessageBubble';
 export { MessageInput } from './MessageInput';
 export { MessageList } from './MessageList';
-export { TeamChatMessageInput } from './TeamChatMessageInput';
 export { TeamChatView } from './TeamChatView';
 export * from './types';
